Fixed files after running for new mac

+++ osx-setup-command.sh
@@ -1,8 +1,7 @@
 #!/bin/bash
 ### --- User selected options ---
 read -r -p "Download qBitTorrent? [y/N] " response
-response=${response,,}
-if [[ "$response" =~ ^(yes|y)$ ]]
+if [[ "$response" =~ ^([yY][eE][sS]|[yY])$ ]]
 then
     download_torrent=true
 else
@@ -10,8 +9,7 @@ else
 fi
 
 read -r -p "Download VMWare Fusion? [y/N] " response
-response=${response,,}
-if [[ "$response" =~ ^(yes|y)$ ]]
+if [[ "$response" =~ ^([yY][eE][sS]|[yY])$ ]]
 then
     download_vmware=true
 else
@@ -19,26 +17,25 @@ else
 fi
 
 ### --- Utilities ---
-/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
 brew update
 brew install \
     git \
     gradle \
     maven \
     python \
-    openjdk@11
+    openjdk@11 \
+    gradle \
+    maven
 
 brew install --cask \
     alfred \
     cheatsheet \
     grammarly \
-    rectangle \
-    gradle \
-    maven
-
+    rectangle 
+    
 ### --- Terminal Customisation ---
-brew install --cask iterm2
 sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+brew install --cask iterm2
 brew install \
     cowsay \
     fortune \
@@ -61,9 +58,8 @@ brew install --cask \
     discord \
     whatsapp \
     appcleaner \
-    bitwarden \
     fork \
-    visual-studio-code
+    visual-studio-code 
 
 if "$download_vmware"; then
     brew install --cask vmware-fusion
@@ -75,11 +71,20 @@ fi
 # Ability to download from mac app store throught CLI
 brew install mas-cli/tap/mas
 mas lucky trello
+# Need to use app store for web auto-comp functionality
+mas lucky bitwarden
+mas lucky onenote 
 
 ### --- MISC ---
 # Use this folder for all new projects
-mkdir ~/dev
-cp .zshrc ~/.zshrc
+sudo mkdir ~/dev
+sudo cp .zshrc ~/.zshrc
+# Show all hideen folders
+defaults write com.apple.Finder AppleShowAllFiles true
 
 # Delete this directory once setup is complete
-rm -rf -- "$(pwd -P)" && cd ..
+read -r -p "Delete this setup directory[y/N] " response
+if [[ "$response" =~ ^([yY][eE][sS]|[yY])$ ]]
+then
+    rm -rf -- "$(pwd -P)" && cd ..
+fi
